requirements.txt
streamlit
pytesseract
openai
python-dotenv
Pillow
